requirements.txt
absl-py==1.1.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
decorator==5.1.1
docker-pycreds==0.4.0
executing==0.8.3
filelock==3.4.0
flake8==4.0.1
flatbuffers==1.12
fvcore==0.1.5.post20220512
gast==0.4.0
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.6.6
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.46.3
h5py==3.7.0
hjson==3.0.2
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.11.4
iopath==0.1.9
ipython==8.4.0
isodate==0.6.0
jedi==0.18.1
jinja2==3.0.3
joblib==1.1.0
keras-bert==0.88.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
keras-preprocessing==1.1.2
keras-self-attention==0.51.0
keras-transformer==0.40.0
keras==2.9.0
libclang==14.0.1
markdown==3.3.7
markupsafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
networkx==2.6.3
ninja==1.10.2.3
numpy==1.21.4
oauthlib==3.2.0
onnx==1.12.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
pip==21.2.4
portalocker==2.4.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptflops==0.6.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
pygments==2.12.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyyaml==6.0
rdflib==6.0.2
regex==2021.11.10
requests-oauthlib==1.3.1
requests==2.26.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
sentencepiece==0.1.96
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==58.0.4
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.2.0
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.0
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
thop==0.1.0.post2206102148
threadpoolctl==3.0.0
tokenizers==0.10.3
torchaudio==0.7.2
tqdm==4.62.3
traitlets==5.2.2.post1
transformers==4.12.5
typing-extensions==4.0.0
urllib3==1.26.7
wandb==0.12.17
wcwidth==0.2.5
werkzeug==2.1.2
wheel==0.37.0
wrapt==1.14.1
yacs==0.1.8
zipp==3.8.0
deepspeed==0.6.5